feat:refactorUserCode
